Introduce module Timeline
Status had methods to select statuses included in timeline. However, there is
also a class named Feed to get statuses in timeline (especially home
timeline). Timeline is a unified module to query statuses in timeline.
