+++ app/services/precompute_feed_service.rb
@@ -34,7 +34,7 @@ class PrecomputeFeedService < BaseService
   end
 
   def statuses
-    Status.as_home_timeline(account).order(account_id: :desc).limit(LIMIT)
+    Timeline.home_from_database(account, limit: LIMIT)
   end
 
   def redis
