+++ spec/services/fan_out_on_write_service_spec.rb
@@ -19,12 +19,12 @@ RSpec.describe FanOutOnWriteService do
   end
 
   it 'delivers status to home timeline' do
-    expect(Feed.new(:home, author).get(10).map(&:id)).to include status.id
+    expect(Timeline.home(author).map(&:id)).to include status.id
   end
 
   it 'delivers status to local followers' do
     pending 'some sort of problem in test environment causes this to sometimes fail'
-    expect(Feed.new(:home, follower).get(10).map(&:id)).to include status.id
+    expect(Timeline.home(follower).map(&:id)).to include status.id
   end
 
   it 'delivers status to hashtag' do
@@ -32,6 +32,6 @@ RSpec.describe FanOutOnWriteService do
   end
 
   it 'delivers status to public timeline' do
-    expect(Status.as_public_timeline(alice).map(&:id)).to include status.id
+    expect(Timeline.public(account: alice).map(&:id)).to include status.id
   end
 end
