+++ spec/services/remove_status_service_spec.rb
@@ -25,11 +25,11 @@ RSpec.describe RemoveStatusService do
   end
 
   it 'removes status from author\'s home feed' do
-    expect(Feed.new(:home, alice).get(10)).to_not include(@status.id)
+    expect(Timeline.home(alice)).to_not include(@status.id)
   end
 
   it 'removes status from local follower\'s home feed' do
-    expect(Feed.new(:home, jeff).get(10)).to_not include(@status.id)
+    expect(Timeline.home(jeff)).to_not include(@status.id)
   end
 
   it 'sends PuSH update to PuSH subscribers' do
